requirements.txt
aioredis==2.0.1
aiosmtplib==1.1.6
anyio==3.6.2
asgiref==3.5.2
async-timeout==4.0.2
autopep8==1.6.0
bcrypt==4.0.1
blinker==1.4
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cryptography==3.4.8
Deprecated==1.2.13
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
fakeredis==1.8.1
fastapi==0.87.0
fastapi-jwt-auth==0.5.0
fastapi-mail==1.0.9
h11==0.14.0
httpcore==0.15.0
httptools==0.4.0
httpx==0.23.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
orjson==3.7.7
packaging==21.3
passlib==1.7.4
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.10.2
PyJWT==1.7.1
pymongo==4.3.3
pyparsing==3.0.9
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
redis==4.3.4
requests==2.28.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
starlette==0.21.0
toml==0.10.2
typing_extensions==4.4.0
ujson==5.4.0
urllib3==1.26.10
uvicorn==0.19.0
watchgod==0.8.2
websockets==10.3
wrapt==1.14.1
